Transfered the app.test code to products.test
